Moved `org.xhtmlrenderer.simple.PDFRenderer` to `org.xhtmlrenderer.pdf.PDFRenderer`
Changed to fix issue https://github.com/flyingsaucerproject/flyingsaucer/issues/349
