+++ www/guide.bat
@@ -1,3 +1,3 @@
 del users-guide-r7.pdf
-java -cp ..\..\build\classes;..\..\lib\cssparser-0-9-4-fs.jar;..\..\lib\itext-paulo-155.jar org.xhtmlrenderer.simple.PDFRenderer d:\java\javanet\xhtmlrenderer\www\r7\users-guide-r7.html users-guide-r7.pdf
+java -cp ..\..\build\classes;..\..\lib\cssparser-0-9-4-fs.jar;..\..\lib\itext-paulo-155.jar org.xhtmlrenderer.pdf.PDFRenderer d:\java\javanet\xhtmlrenderer\www\r7\users-guide-r7.html users-guide-r7.pdf
 start users-guide-r7.pdf
